flask_api/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
pystray==0.19.3
Pillow==10.3.0
werkzeug==3.0.1
PyQt5==5.15.11
pyinstaller==6.14.1
